reorganized the backend/cleaned up
