+++ server/models/eventsModel.js
@@ -1,0 +1,17 @@
+const mongoose = require('mongoose');
+
+const eventSchema = new mongoose.Schema({
+
+    title: { 
+        type: String, required: true 
+    },
+    photo: {
+        type: String, required: true 
+    },
+    date: { 
+        type: Date, required: true 
+    },
+    
+});
+
+module.exports = mongoose.model('Event', eventSchema);
