+++ server/routes/goalRoutes.js
@@ -1,15 +1,10 @@
 const express = require("express");
 const router = express.Router();
-const {
-    addGoal,
-    editGoal,
-    getGoals,
-    deleteGoal,
-  } = require('../controllers/users/goalController');
+const {addGoal, editGoal, getGoals, deleteGoal,} = require('../controllers/users/goal.controllers.js');
 
-  router.post("/addGoal", addGoal);
-  router.get("/getGoal/:userId", getGoals);
-  router.put("/editGoal/:id", editGoal);
-  router.delete("/deleteGoal/:id", deleteGoal);
+router.post("/addGoal", addGoal);
+router.get("/getGoal/:userId", getGoals);
+router.put("/editGoal/:id", editGoal);
+router.delete("/deleteGoal/:id", deleteGoal);
 
-  module.exports = router;
+module.exports = router;
